Split InAir into Falling and Jumping
