+++ script/player/States/Idle.cs
@@ -13,15 +13,18 @@ namespace INTOnlineCoop.Script.Player.States
         /// <param name="delta">Current frame delta</param>
         public override void PhysicProcess(double delta)
         {
+            if (!Character.IsOnFloor())
+            {
+                Character.StateMachine.TransitionTo(AvailableState.Falling);
+            }
+            else if (Input.IsActionJustPressed("jump"))
+            {
+                Character.StateMachine.TransitionTo(AvailableState.Jumping);
+            }
             //Übergang in den Walking-Zustand, falls Eingabe erfolgt
-            if (Input.IsActionPressed("walk_right") || Input.IsActionPressed("walk_left"))
+            else if (Input.IsActionPressed("walk_right") || Input.IsActionPressed("walk_left"))
             {
                 Character.StateMachine.TransitionTo(AvailableState.Walk);
-            }
-            //Übergang in den Inair-Zustand, falls in der Luft oder Sprung
-            else if (!Character.IsOnFloor() || Input.IsActionJustPressed("jump"))
-            {
-                Character.StateMachine.TransitionTo(AvailableState.InAir);
             }
         }
     }
